requirements.txt
# requirements.txt
pyTelegramBotAPI==4.15.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.8.5
flask==2.3.3
gunicorn==21.2.0